reorganised the project into modules
